Refactor CLI evidence bundle commands
